examples/crewai_agent/requirements.txt
bunkervm[crewai]
python-dotenv
